Initial check-in.
Initial check-in of a basically working Jam language parser. A few
language features are still missing, like the maxline actions flag.
The parser generates a tree of nodes that can be executed (though
that is mostly not yet implemented) and dumped. Compilation of the
grammar source file is horribly slow.