testdata/make/Update: Compatibility handling for test case 22
While Jam considers targets not marked NotFile that are missing,
don't have actions, and have dependencies pseudo targets, it doesn't do
so consequently. A target depending on such a target is still updated,
even when up-to-date. Ham doesn't do that. Strictly speaking it would be
better to fail in such a case, since it really is a user error not to
mark the target a pseudo target and probably just complicates trouble
shooting for them.